refactor: move network tokens module
