+++ src/frontend/src/lib/components/exchange/ExchangeBalance.svelte
@@ -3,7 +3,7 @@
 	import { exchangeInitialized, exchanges } from '$lib/derived/exchange.derived';
 	import { balancesStore } from '$lib/stores/balances.store';
 	import { usdValue } from '$lib/utils/exchange.utils';
-	import { networkTokens } from '$lib/derived/network.derived';
+	import { networkTokens } from '$lib/derived/network-tokens.derived';
 
 	let totalUsd: number;
 	$: totalUsd = $networkTokens.reduce(
