+++ src/frontend/src/lib/derived/network.derived.ts
@@ -3,15 +3,9 @@ import { DEFAULT_NETWORK, DEFAULT_NETWORK_ID } from '$lib/constants/networks.con
 import { address } from '$lib/derived/address.derived';
 import { routeNetwork } from '$lib/derived/nav.derived';
 import { networks } from '$lib/derived/networks.derived';
-import { tokens } from '$lib/derived/tokens.derived';
 import type { OptionAddress } from '$lib/types/address';
 import type { Network, NetworkId } from '$lib/types/network';
-import type { Token } from '$lib/types/token';
-import {
-	isNetworkIdChainFusion,
-	isNetworkIdEthereum,
-	isNetworkIdICP
-} from '$lib/utils/network.utils';
+import { isNetworkIdEthereum, isNetworkIdICP } from '$lib/utils/network.utils';
 import { nonNullish } from '@dfinity/utils';
 import { derived, type Readable } from 'svelte/store';
 
@@ -26,17 +20,6 @@ export const networkId: Readable<NetworkId> = derived(
 export const selectedNetwork: Readable<Network> = derived(
 	[networks, networkId],
 	([$networks, $networkId]) => $networks.find(({ id }) => id === $networkId) ?? DEFAULT_NETWORK
-);
-
-// TODO: move me to a separate module to express the dependency on tokens
-export const networkTokens: Readable<Token[]> = derived(
-	[tokens, selectedNetwork],
-	([$tokens, $selectedNetwork]) =>
-		$tokens.filter(
-			({ network: { id, env } }) =>
-				(isNetworkIdChainFusion($selectedNetwork.id) && env === $selectedNetwork.env) ||
-				id === $selectedNetwork.id
-		)
 );
 
 export const networkICP: Readable<boolean> = derived([networkId], ([$networkId]) =>
